[backend] api: introduce headerbar api
