--- backend/public/css/style.css
@@ -1,3 +1,0 @@
-.site-logo {
-  max-height: 40px;
-}
